+++ src/util/constants.js
@@ -1,3 +1,7 @@
 // eslint-disable-next-line import/prefer-default-export
 export const MAX_FAILED_ATTEMPTS = 5;
 export const DAYS_TO_KEEP_LOGS = 60;
+
+export const CS4 = 'cs4';
+export const CS50 = 'cs50';
+export const CS51 = 'cs51';