+++ test/jest/setupFiles.js
@@ -1,1 +1,5 @@
 import './__mock__';
+
+import { TextDecoder, TextEncoder } from 'util';
+global.TextEncoder = TextEncoder;
+global.TextDecoder = TextDecoder;